Add models
I added the necessary database models along with the respective relations,
validations and migrations. I already added methods to book returns and
lendings to the Student model. Lendings is polymorphic: the lender may be
a student or a teacher. Furthermore, every model with a has_many relationship
is set to delete dependent data sets.