Complete school actions
Schools can now be created (using the signup page), edited (change password),
and deleted. Therefore, all school actions are implemented. The latter two
options are only available if you are logged in and you can obviously only
delete and edit your own school account. Apart from that, I implemented logging
in and out and the correct display of the respective actions in the menu.