Add lending functionality
I implemented lending and returning of books (base set lendings as well as
miscellaneous lendings). When an isbn field loses focus, the corresponding
book will be looked up (via an Ajax request). For the lookup the aliases table
will be queried first and if the input value is not an alias, it will be treated
as an isbn and if that doesn't yield a result either, an error message will be
displayed in red. Furthermore, the controller now properly differenciates
between base set lendings and miscellaneous lendings.