Add teachers controller
The teachers controller will handle all actions related to teachers. Until now,
I only added the index action which displays all of current_school's teachers
(not unlike students#index) with options for deletion and editing. Moreover, I
tweaked the students#index view so that the options for insertion and import are
no longer plain buttons with the respective captions but green buttons with
glyphicons which are positioned next to the title.