Complete teachers controller
Apart from adding the missing RESTful actions to the teachers controller, I
tweaked the students' index page so that it also contains a list of classes
which is currently located above the students list. Besides, when invoking
students#new from the class's page, the graduation year and the class letter
are set automatically. The same will happen with the students#import action
when I implement it.