Change database to store a book's form in the usages table
Schools use books differently. For example one school might use Adeo norm for
Latin in 9th and 10th form, while the other also uses it for 11th and 12th form.
This makes it necessary to move the form column from the books table to the
usages table, which I did in this commit.