Add friendly forwarding
After lending or returning books, the user will now be redirected to the page
he was on beforehand (e.g. students#show_class).
Apart from that, I fixed the bug where in the students#index view, the 11th
class was displayed multiple times (with my seed data). It turns out, that the
seed data sets class_letter to NULL while an empty text field sets it to an
empty string. Naturally, I corrected the seed data.