Implement import functionality
When pressing the import button in the students#index or students#show_class
views, a form will be retrieved (via AJAX) that allows you to select a CSV file,
offers fields for common values (such as graduation year or class letter) and
provides help as to how these CSV files are to be structured. When clicking
import, the students will be imported and a success message or errors will be
displayed.