Add aliases controller
To manage all aliases, I added the aliases controller with #index, #new, #create
and #destroy actions. Aliases can be created from the aliases#index
or from the books#show page where a form can be loaded via AJAX. The books#show
page shows (apart from the title) the students with have lent the book (as
miscellaneous lending) and the classes to which it was provided as base set.