Complete search field
The queries are looked up using simply the LIKE operator. To accelerate this
process, I added indices to the respective columns. Furthermore, I added
styling so that the suggestions are no longer transparent. When you click on a
suggestion, you are automatically directed to the respective #show view.
Apart from that, I deleted the typeahead.js file and replaced it with the
bootstrap-typeahead-rails gem.