Add differentiation between new and old list in students#show_class
I added a select box to the students#show_class view that lets the user choose
whether the old or the new list should be displayed (the old books and the new
books). Selection of an item leads to an AJAX request that reloads the table.
To accomplish this, I outsourced the table content into a partial, which is also
used to respond to the AJAX request. To ensure further functionality of the
select box, the event handler "change" has to be set again (I pulled this into
a method of its own).