Refactor helper methods
A few methods were defined in helper modules but not used to facilitate the display
of data. Instead they were primarily used in controllers, so I moved those methods
to the ApplicationController or a respective data model. The methods used to display
data were, of course, left in place. Several methods were used in views as well as in
controllers, those I moved to the ApplicationController and declared as helper_method.
The result is that the ApplicationController doesn't depend on (i.e. include) any
helpers anymore.