Various tweaks with regard to style
I changed the color of (and added a gradient to) the navbar. Besides, I changed the
static#home view to only show the jumbotron unless the user is logged in, when he/she is
logged in, buttons which lead to the respective tables are displayed. Furthermore, the
jumbotron now includes a log in button.