Adjust pages to smaller laptop displays
The navbar was becoming very crowded so that on smaller laptop displays a line break was
inserted, which rendered the heading of the page illegible. I fixed this by pushing all
links related to tables into a dropdown ("Tabellen"). Furthermore, I changed the caption
of the mass edit button to "Bearbeiten" instead of "Ändern".