Change books to be school-specific
A book now belongs to a school that uses it. This is due to a book record not
standing for a specific book (e.g. "Ikarus Physik 7") but rather for a role
(i.e. physics textbook for 7th class) so that it is possible to swap books
to a new edition or different publisher (so that the ISBN changes) without the
need to change every `lending` or `base_set` record ever created for that book.
You just change the ISBN and the title and no other schools are affected.

I added two migrations: one for adding fields, foreign_keys, indices, etc. to
the `books` table, and one for dropping the `usages` table. As a consequence,
I had to modify the `Book` model and the `School` model (`has_many :books`
instead of `... :through => :usages`).
Finally, I removed `Book#display_title`, `BooksHelper#display_title`,
`ApplicationController#display_title`, as well as `ApplicationController#form`
because they are no longer needed due to the fact that henceforth only the
`Book`'s `title` and only the `Book`'s `title` shall be displayed.