Change key format back to dasherised
In order to comply fully with the JSONAPI and to minimise the amount of
customisation necessary on the client side, I changed the format for model,
attributes, relationships, links, etc. to dasherised. This means that
the initializer for jsonapi-resources is superfluous. Furthermore, I adjusted
the tests to the new naming scheme and updated the README.

When creating new teachers, students, books or schools, 422 (or 500) was
returned. The reason being, that the before_create callback checked for the
presence of 'school' in the model class's instance methods, which in itself is
an array of symbols. 500 was returned when trying to create a new school. This
was due to ApplicationController#context using the @@current_school class
variable which hadn't yet been initialised (it is usually initialised in
authenticate_from_token! which is skipped during schools#create).